cart service methods and functionality added

+++ coupon-service/src/main/java/smr/shop/coupon/service/model/CouponEntity.java
@@ -5,9 +5,10 @@ import jakarta.persistence.EnumType;
 import jakarta.persistence.Enumerated;
 import jakarta.persistence.Id;
 import lombok.*;
-import smr.shop.coupon.service.model.valueObject.CouponDiscountType;
-import smr.shop.coupon.service.model.valueObject.CouponType;
+import smr.shop.coupon.service.model.valueobject.CouponDiscountType;
+import smr.shop.coupon.service.model.valueobject.CouponType;
 
+import java.math.BigDecimal;
 import java.time.ZonedDateTime;
 import java.util.UUID;
 
@@ -33,11 +34,11 @@ public class CouponEntity {
     @Enumerated(EnumType.STRING)
     private CouponDiscountType discountType;
 
-    private Double amount;
+    private BigDecimal amount;
 
-    private Float percentage;
+    private Short percentage;
 
-    private Double maxDiscountPrice;
+    private BigDecimal maxDiscountPrice;
 
     private ZonedDateTime endDate;
